+++ src/cosmos/cosmosTypes.ts
@@ -70,14 +70,23 @@ export interface CosmosNetworkInfo {
   bech32AddressPrefix: string
   bip39Path: string
   chainInfo: {
-    chainId: string
+    chainName: string
     url: string
   }
+  defaultChainId: string
+  chainIdUpdateUrl?: string
   defaultTransactionFeeUrl?: HttpEndpoint
   nativeDenom: string
   pluginMnemonicKeyName: string
   rpcNode: HttpEndpoint
-  archiveNode?: HttpEndpoint // If no archive node, the rpc node will be used and only grab transaction in a limited range
+  // If no archive node, the rpc node will be used and only grab transaction in a limited range
+  archiveNodes?: Array<{
+    blockTimeRangeSeconds: {
+      start: number
+      end?: number
+    }
+    endpoint: HttpEndpoint
+  }>
 }
 const asHttpEndpoint = asObject({
   url: asString,
@@ -90,19 +99,16 @@ export const txQueryStrings = [
 ] as const
 
 const asLocalTxQueryParams = asObject({
-  newestTxid: asMaybe(asString),
-  lastPage: asMaybe(asNumber, 1)
+  newestTxid: asMaybe(asString)
 })
 
 export const asCosmosWalletOtherData = asObject({
   archivedTxLastCheckTime: asMaybe(asNumber, 0),
   'coin_spent.spender': asMaybe(asLocalTxQueryParams, () => ({
-    newestTxid: undefined,
-    lastPage: 1
+    newestTxid: undefined
   })),
   'coin_received.receiver': asMaybe(asLocalTxQueryParams, () => ({
-    newestTxid: undefined,
-    lastPage: 1
+    newestTxid: undefined
   }))
 })
 export type CosmosWalletOtherData = ReturnType<typeof asCosmosWalletOtherData>
@@ -272,6 +278,14 @@ export interface IbcChannel {
 
 export const asCosmosInfoPayload = asObject({
   rpcNode: asOptional(asHttpEndpoint),
-  archiveNode: asOptional(asHttpEndpoint)
+  archiveNodes: asOptional(asHttpEndpoint)
 })
 export type CosmosInfoPayload = ReturnType<typeof asCosmosInfoPayload>
+
+export const asChainIdUpdate = asObject({
+  result: asObject({
+    node_info: asObject({
+      network: asString // 'thorchain-mainnet-v1',
+    })
+  })
+})