+++ src/cosmos/info/axelarInfo.ts
@@ -12,19 +12,27 @@ const networkInfo: CosmosNetworkInfo = {
   bech32AddressPrefix: 'axelar',
   bip39Path: `m/44'/118'/0'/0/0`,
   chainInfo: {
-    chainId: 'axelar-dojo-1',
+    chainName: 'axelar',
     url: 'https://raw.githubusercontent.com/cosmos/chain-registry/master/axelar/chain.json'
   },
+  defaultChainId: 'axelar-dojo-1',
   nativeDenom: 'uaxl',
   pluginMnemonicKeyName: 'axelarMnemonic',
   rpcNode: {
     url: 'https://axelar-rpc.publicnode.com:443',
     headers: {}
   },
-  archiveNode: {
-    url: 'https://axelararchive-rpc.quickapi.com:443',
-    headers: {}
-  }
+  archiveNodes: [
+    {
+      blockTimeRangeSeconds: {
+        start: 0
+      },
+      endpoint: {
+        url: 'https://axelararchive-rpc.quickapi.com:443',
+        headers: {}
+      }
+    }
+  ]
 }
 
 const currencyInfo: EdgeCurrencyInfo = {